+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## Unreleased
+- Fix lazyness bug in `discover-network`: was sending multiple `Who-Is` simultaneously.
 
 ## [1.2.0] - 2024-01-21
 - `with-temp-devices` no longer terminates and restarts the local device.